Add e2-demo-report and python conversion script
A sample report from DETRV to FIHEL was received and multiple waypoints
were inserted in the middle of the route, replacing some waypoints. To
speed up the process of making sample routes in the future a Python
script was written to give the waypoints new ids and to add the same
amount of waypoints to the schedule.